# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API to manage employees and departments of a company.",
        "title": "Employee and Department Management API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/departamentos": {
            "post": {
                "description": "Creates a new department with the provided data",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Departamentos"
                ],
                "summary": "Create a new department",
                "parameters": [
                    {
                        "description": "Department data",
                        "name": "department",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CreateDepartmentDTO"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.Department"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error (Invalid Manager/Parent Department)",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/departamentos/listar": {
            "post": {
                "description": "Returns a paginated list of departments based on filters",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Departamentos"
                ],
                "summary": "List departments with filters",
                "parameters": [
                    {
                        "description": "Filters and pagination",
                        "name": "filters",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/models.ListDepartmentsDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Department"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/departamentos/{id}": {
            "get": {
                "description": "Returns department with manager and complete hierarchical tree of sub-departments",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Departamentos"
                ],
                "summary": "Get department by ID with hierarchical tree",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Department ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Department with SubDepartments filled",
                        "schema": {
                            "$ref": "#/definitions/models.Department"
                        }
                    },
                    "400": {
                        "description": "Invalid ID",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Department not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "put": {
                "description": "Updates department data and prevents cycles",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Departamentos"
                ],
                "summary": "Update a department",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Department ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Data to update",
                        "name": "department",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.UpdateDepartmentDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Department"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Department not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error (Invalid Manager/Parent Department or Cycle detected)",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "description": "Removes a department (soft delete)",
                "tags": [
                    "Departamentos"
                ],
                "summary": "Delete a department",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Department ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Invalid ID",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Department not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "422": {
                        "description": "Cannot remove department with employees or sub-departments",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "models.CreateDepartmentDTO": {
            "type": "object",
            "required": [
                "manager_id",
                "name"
            ],
            "properties": {
                "manager_id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "parent_department_id": {
                    "type": "string"
                }
            }
        },
...
        "models.Department": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "manager": {
                    "$ref": "#/definitions/models.Employee"
                },
                "manager_id": {
                    "description": "Pointer to accept NULL",
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "parent_department_id": {
                    "description": "Pointer to accept NULL",
                    "type": "string"
                },
                "sub_departments": {
                    "description": "Used to load the hierarchical tree",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Department"
                    }
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "models.Employee": {
            "type": "object",
            "properties": {
                "cpf": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "department_id": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "rg": {
                    "description": "Pointer to accept NULL",
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "models.ListDepartmentsDTO": {
            "type": "object",
            "properties": {
                "manager_name": {
                    "description": "Special filter",
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "page": {
                    "type": "integer",
                    "minimum": 1
                },
                "page_size": {
                    "type": "integer",
                    "minimum": 1
                },
                "parent_department_id": {
                    "type": "string"
                }
            }
        },
...
        "models.UpdateDepartmentDTO": {
            "type": "object",
            "properties": {
                "manager_id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "parent_department_id": {
                    "type": "string"
                }
            }
        },
...
    }
}